refactor mozart
