+++ .bruno/Exercises/CreateExerciseHaskell.bru
@@ -21,7 +21,7 @@ body:json {
     "Solution": "module Solution where\nsolution x =\n  if x < 0\n    then x * (-1)\n    else x",
     "solutionLanguage": 1,
     "InputParameterType": ["int"],
-    "OutputParamaterType": ["int"],
+    "OutputParameterType": ["int"],
     "Testcases": [
       {
         "inputParams": ["-5"],
