+++ .bruno/Exercises/CreateExercisePython.bru
@@ -21,7 +21,7 @@ body:json {
     "Solution": "def solution(x: int):\n    return x + 1",
     "solutionLanguage": 2,
     "InputParameterType": ["int"],
-    "OutputParamaterType": ["int"],
+    "OutputParameterType": ["int"],
     "Testcases": [
       {
         "inputParams": ["5"],
